requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
djangorestframework==3.14.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.2
